buffixes

+++ src/org/puzzle/ui/Drawable.java
@@ -11,10 +11,10 @@ import android.graphics.Paint;
 public interface Drawable {
 
 	
-	public int getX();
-	public int getY();
-	public void setX(int x);
-	public void setY(int y);
+	public float getX();
+	public float getY();
+	public void setX(float x);
+	public void setY(float y);
 
 	
 	/**
